# [LinearSolver.Direct] LU solver can be templated to CRSMat3x3