--- a/Component/LinearSolver/Direct/src/sofa/component/linearsolver/direct/SparseLUSolver.h
+++ b/Component/LinearSolver/Direct/src/sofa/component/linearsolver/direct/SparseLUSolver.h
@@ -89,6 +89,8 @@ protected :
         return new SparseLUInvertData<Real>();
     }
 
+    sofa::linearalgebra::CompressedRowSparseMatrix<Real> Mfiltered;
+
 };