--- a/Component/LinearSolver/Direct/src/sofa/component/linearsolver/direct/SparseLUSolver.h
+++ b/Component/LinearSolver/Direct/src/sofa/component/linearsolver/direct/SparseLUSolver.h
@@ -89,7 +89,7 @@ protected :
         return new SparseLUInvertData<Real>();
     }
 
-    sofa::linearalgebra::CompressedRowSparseMatrix<Real> Mfiltered;
+    std::unique_ptr<sofa::linearalgebra::CompressedRowSparseMatrix<Real> > Mfiltered;
 
 };