--- a/Component/LinearSolver/Direct/src/sofa/component/linearsolver/direct/SparseLUSolver.inl
+++ b/Component/LinearSolver/Direct/src/sofa/component/linearsolver/direct/SparseLUSolver.inl
@@ -98,21 +98,37 @@ void SparseLUSolver<TMatrix,TVector,TThreadManager>::invert(Matrix& M)
 {
     SparseLUInvertData<Real> * invertData = (SparseLUInvertData<Real>*) this->getMatrixInvertData(&M);
 
-    Mfiltered.copyNonZeros(M);
-    Mfiltered.compress();
+    sofa::linearalgebra::CompressedRowSparseMatrix<Real>* matrix;
+
+    if constexpr (!std::is_same_v<Matrix, decltype(Mfiltered)>)
+    {
+        if (!Mfiltered)
+        {
+            Mfiltered = std::make_unique<sofa::linearalgebra::CompressedRowSparseMatrix<Real> >();
+        }
+        Mfiltered->copyNonZeros(M);
+        Mfiltered->compress();
+
+        matrix = Mfiltered.get();
+    }
+    else
+    {
+        M.compress();
+        matrix = &M;
+    }
 
     if (invertData->N) cs_nfree(invertData->N);
     if (invertData->tmp) cs_free(invertData->tmp);
 
     //build A with M
-    invertData->A.nzmax = Mfiltered.getColsValue().size();	// maximum number of entries
-    invertData->A.m = Mfiltered.rowSize();					// number of rows
-    invertData->A.n = Mfiltered.colSize();					// number of columns
-    invertData->A_p = Mfiltered.getRowBegin();
+    invertData->A.nzmax = matrix->getColsValue().size();	// maximum number of entries
+    invertData->A.m = matrix->rowSize();					// number of rows
+    invertData->A.n = matrix->colSize();					// number of columns
+    invertData->A_p = matrix->getRowBegin();
     invertData->A.p = (int *) &(invertData->A_p[0]);							// column pointers (size n+1) or col indices (size nzmax)
-    invertData->A_i = Mfiltered.getColsIndex();
+    invertData->A_i = matrix->getColsIndex();
     invertData->A.i = (int *) &(invertData->A_i[0]);							// row indices, size nzmax
-    invertData->A_x = Mfiltered.getColsValue();
+    invertData->A_x = matrix->getColsValue();
     invertData->A.x = (Real *) &(invertData->A_x[0]);				// numerical values, size nzmax
     invertData->A.nz = -1;							// # of entries in triplet matrix, -1 for compressed-col
     cs_dropzeros( &invertData->A );