--- a/examples/Components/linearsolver/FEMBAR_SparseLUSolver.scn
+++ b/examples/Components/linearsolver/FEMBAR_SparseLUSolver.scn
@@ -2,7 +2,7 @@
 
     <include href="FEMBAR-common.xml"/>
 
-    <SparseLUSolver/>
+    <SparseLUSolver template="CompressedRowSparseMatrixMat3x3d" permutation="None"/>
     <HexahedronFEMForceField name="FEM" youngModulus="4000" poissonRatio="0.3" method="large" />
 
 </Node>