Matrix assembly can be accelerated if CRS matrices are defined using blocks of 3x3.
This PR makes LU solver compatible with 3x3 blocks.

Benchmark https://github.com/alxbilger/SofaBenchmark/pull/22:

Before

```
---------------------------------------------------------------------------------------------------
Benchmark                                         Time             CPU   Iterations UserCounters...
---------------------------------------------------------------------------------------------------
BM_SparseLUSolver/50/iterations:10              368 ms          370 ms           10 FPS=135.021/s MBKBuild=1.82966m MBKSolve=5.34694m frame=7.40625ms
```

After

```
---------------------------------------------------------------------------------------------------
Benchmark                                         Time             CPU   Iterations UserCounters...
---------------------------------------------------------------------------------------------------
BM_SparseLUSolver/50/iterations:10              371 ms          367 ms           10 FPS=136.17/s MBKBuild=1.81682m MBKSolve=5.41681m frame=7.34375ms
BM_SparseLUSolverMat3x3/50/iterations:10        308 ms          311 ms           10 FPS=160.804/s MBKBuild=529.916u MBKSolve=5.4368m frame=6.21875ms
```

We can observe that the time stays the same for `CompressedRowSparseMatrixd` (no performance loss), and it is faster (MBKBuild timer) for `CompressedRowSparseMatrixd`.


______________________________________________________

By submitting this pull request, I acknowledge that  
**I have read, understand, and agree [SOFA Developer Certificate of Origin (DCO)](https://github.com/sofa-framework/sofa/blob/master/CONTRIBUTING.md#sofa-developer-certificate-of-origin-dco)**.
______________________________________________________

**Reviewers will merge this pull-request only if**  
- it builds with SUCCESS for all platforms on the CI.
- it does not generate new warnings.
- it does not generate new unit test failures.
- it does not generate new scene test failures.
- it does not break API compatibility.
- it is more than 1 week old (or has fast-merge label).